Embrace more SOLID principles
