+++ plugins-internal/src/test/java/dev/cookiecode/rika2mqtt/plugins/internal/v1/HttpPluginDownloaderTest.java
@@ -1,0 +1,62 @@
+/*
+ * The MIT License
+ * Copyright © 2022 Sebastien Vermeille
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+package dev.cookiecode.rika2mqtt.plugins.internal.v1;
+
+import static java.net.URI.create;
+import static org.assertj.core.api.Assertions.assertThat;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.io.TempDir;
+import org.mockito.InjectMocks;
+import org.mockito.Spy;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+/**
+ * Test class
+ *
+ * @author Sebastien Vermeille
+ */
+@ExtendWith(MockitoExtension.class)
+class HttpPluginDownloaderTest {
+
+  @TempDir public Path pluginDir;
+
+  @InjectMocks @Spy private HttpPluginDownloader httpPluginDownloader;
+
+  @Test
+  void downloadPluginShouldDownloadTheFileGivenItExists() throws Exception {
+    // GIVEN
+    final var pluginUrl = "https://repo.maven.apache.org/maven2/dev/cookiecode/maven-metadata.xml";
+
+    // WHEN
+    httpPluginDownloader.downloadPlugin(
+        create(pluginUrl).toURL(), pluginDir.toAbsolutePath().toString());
+
+    // THEN
+    final var expectedFile = Paths.get(pluginDir.toAbsolutePath().toString(), "maven-metadata.xml");
+    assertThat(Files.exists(expectedFile)).as("File has been downloaded").isTrue();
+  }
+}
